// src/models/user.ts
export interface User {
    utility_bill: any;
    title: string;
    first_name: string;
    last_name: string;
    phone: string;
    address: string;
    country_id: number;
    birthday: string;
    status: string;
    other_names: string;
    mobile: string;
    address_1: string;
    address_2: string;
    plan_to_invest: string;
    signup_type: number;
    town: string;
    post_code: string;
    d_country: string;
    nationality: string;
    dob: string;
    marital_status: string;
    dependants_number: string;
    gender: string;
    uk_ni_number: string;
    green_card_holder: string;
    tax_country: string;
    employment_type: string;
    employer: string;
    nature_of_business: string;
    occupation: string;
    employer_street_1: string;
    employer_street_2: string;
    employer_city: string;
    employer_state_province: string;
    employer_zip_postal_code: string;
    employer_country: string;
    net_worth: string;
    liquid_net_income: string;
    annual_net_income: string;
    total_assets: string;
    specific_goal: string;
    risk_level: string;
    annual_goal: string;
    modelPortfolio: string;
    tos: string;
}